fix staticmethod error in python 3.9
